flask/requirements.txt
flask
gunicorn
scikit-learn
pandas
flask-oauthlib
flask-sqlalchemy
flask_pymongo